backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
joblib
scikit-learn
pandas
python-dotenv
